fix : 매너온도 표시문제 수정

+++ src/main/java/com/jungle/studybbitback/domain/room/dto/roommember/GetRoomMemberResponseDto.java
@@ -4,6 +4,7 @@ import com.jungle.studybbitback.domain.dailystudy.entity.DailyStudy;
 import com.jungle.studybbitback.domain.room.entity.RoomMember;
 import lombok.Getter;
 
+import java.math.BigDecimal;
 import java.time.Duration;
 
 @Getter
@@ -12,17 +13,8 @@ public class GetRoomMemberResponseDto {
     private String nickname;
     private String leaderLabel;
     private String profileImageUrl;
-    private Double flowTemperature;
+    private BigDecimal flowTemperature;
     private Duration studyTime;
-
-    // 새로운 생성자 추가 (RoomMember와 leaderId를 매개변수로 받음)
-//    public GetRoomMemberResponseDto(RoomMember roomMember, Long leaderId) {
-//        this.memberId = roomMember.getMember().getId();
-//        this.nickname = roomMember.getMember().getNickname();
-//        this.leaderLabel = roomMember.getMember().getId().equals(leaderId) ? "방장" : ""; // 방장이면 "방장", 아니면 빈 문자열
-//        this.profileImageUrl = roomMember.getMember().getProfileImageUrl(); //null이면 기본 이미지
-//        this.flowTemperature = roomMember.getMember().getFlowTemperature();
-//    }
 
 	public GetRoomMemberResponseDto(RoomMember roomMember, DailyStudy dailyStudy, Long leaderId) {
         this.memberId = roomMember.getMember().getId();
