+++ src/main/java/com/jungle/studybbitback/domain/room/dto/schedulemember/GetScheduleMemberResponseDto.java
@@ -3,8 +3,9 @@ package com.jungle.studybbitback.domain.room.dto.schedulemember;
 import com.jungle.studybbitback.domain.room.entity.schedule.ParticipateStatusEnum;
 import com.jungle.studybbitback.domain.room.entity.schedule.ScheduleMember;
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Getter;
+
+import java.math.BigDecimal;
 
 @Getter
 @AllArgsConstructor
@@ -13,7 +14,7 @@ public class GetScheduleMemberResponseDto {
     private String memberNickname;
     private ParticipateStatusEnum participateStatus;
     private String notedDetail;
-    private Double flowTemperature;
+    private BigDecimal flowTemperature;
 
     public static GetScheduleMemberResponseDto from(ScheduleMember scheduleMember) {
         return new GetScheduleMemberResponseDto(
